+++ crates/gpui/src/executor2.rs
@@ -1,12 +1,8 @@
 use crate::App;
-use async_task::Runnable;
 use futures::channel::mpsc;
 use scheduler::Timer;
 use smol::prelude::*;
-use std::mem::ManuallyDrop;
-use std::panic::Location;
 use std::sync::Arc;
-use std::thread::{self, ThreadId};
 use std::{
     fmt::Debug,
     marker::PhantomData,
@@ -123,8 +119,14 @@ impl BackgroundExecutor {
     where
         R: Send + 'static,
     {
-        // todo!("reduce monomorphization")
-        Task(self.0.spawn(future))
+        fn inner<R: Send + 'static>(
+            executor: &scheduler::BackgroundExecutor,
+            future: AnyFuture<R>,
+        ) -> Task<R> {
+            Task(executor.spawn(future))
+        }
+
+        inner(&self.0, Box::pin(future))
     }
 
     /// Enqueues the given future to be run to completion on a background thread.
@@ -308,88 +310,16 @@ impl ForegroundExecutor {
     where
         R: 'static,
     {
-        // let dispatcher = self.dispatcher.clone();
-
-        // #[track_caller]
-        // fn inner<R: 'static>(
-        //     dispatcher: Arc<dyn PlatformDispatcher>,
-        //     future: AnyLocalFuture<R>,
-        // ) -> Task<R> {
-        //     let (runnable, task) = spawn_local_with_source_location(future, move |runnable| {
-        //         dispatcher.dispatch_on_main_thread(runnable)
-        //     });
-        //     runnable.schedule();
-        //     Task(TaskState::Spawned(task))
-        // }
-        // inner::<R>(dispatcher, Box::pin(future))
-        // todo!("reduce monomorphization and spawn_local_with_source_location")
-        Task(self.0.spawn(future))
-    }
-}
-
-/// Variant of `async_task::spawn_local` that includes the source location of the spawn in panics.
-///
-/// Copy-modified from:
-/// <https://github.com/smol-rs/async-task/blob/ca9dbe1db9c422fd765847fa91306e30a6bb58a9/src/runnable.rs#L405>
-#[track_caller]
-fn spawn_local_with_source_location<Fut, S>(
-    future: Fut,
-    schedule: S,
-) -> (Runnable<()>, async_task::Task<Fut::Output, ()>)
-where
-    Fut: Future + 'static,
-    Fut::Output: 'static,
-    S: async_task::Schedule<()> + Send + Sync + 'static,
-{
-    #[inline]
-    fn thread_id() -> ThreadId {
-        std::thread_local! {
-            static ID: ThreadId = thread::current().id();
+        #[track_caller]
+        fn inner<R: 'static>(
+            executor: &scheduler::ForegroundExecutor,
+            future: AnyLocalFuture<R>,
+        ) -> Task<R> {
+            Task(executor.spawn(future))
         }
-        ID.try_with(|id| *id)
-            .unwrap_or_else(|_| thread::current().id())
+
+        inner::<R>(&self.0, Box::pin(future))
     }
-
-    struct Checked<F> {
-        id: ThreadId,
-        inner: ManuallyDrop<F>,
-        location: &'static Location<'static>,
-    }
-
-    impl<F> Drop for Checked<F> {
-        fn drop(&mut self) {
-            assert!(
-                self.id == thread_id(),
-                "local task dropped by a thread that didn't spawn it. Task spawned at {}",
-                self.location
-            );
-            unsafe {
-                ManuallyDrop::drop(&mut self.inner);
-            }
-        }
-    }
-
-    impl<F: Future> Future for Checked<F> {
-        type Output = F::Output;
-
-        fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-            assert!(
-                self.id == thread_id(),
-                "local task polled by a thread that didn't spawn it. Task spawned at {}",
-                self.location
-            );
-            unsafe { self.map_unchecked_mut(|c| &mut *c.inner).poll(cx) }
-        }
-    }
-
-    // Wrap the future into one that checks which thread it's on.
-    let future = Checked {
-        id: thread_id(),
-        inner: ManuallyDrop::new(future),
-        location: Location::caller(),
-    };
-
-    unsafe { async_task::spawn_unchecked(future, schedule) }
 }
 
 /// Scope manages a set of tasks that are enqueued and waited on together. See [`BackgroundExecutor::scoped`].
