+++ crates/gpui/src/platform/mac.rs
@@ -1,10 +1,10 @@
 //! Macos screen have a y axis that goings up from the bottom of the screen and
 //! an origin at the bottom left of the main display.
-mod dispatcher;
 mod display;
 mod display_link;
 mod events;
 mod keyboard;
+mod scheduler;
 
 #[cfg(feature = "screen-capture")]
 mod screen_capture;
@@ -45,11 +45,11 @@ use std::{
     ops::Range,
 };
 
-pub(crate) use dispatcher::*;
 pub(crate) use display::*;
 pub(crate) use display_link::*;
 pub(crate) use keyboard::*;
 pub(crate) use platform::*;
+pub(crate) use scheduler::*;
 pub(crate) use window::*;
 
 #[cfg(feature = "font-kit")]
