+++ crates/gpui/src/platform/test.rs
@@ -1,12 +1,10 @@
-mod dispatcher;
 mod display;
 mod platform;
 mod window;
 
-pub use dispatcher::*;
 pub(crate) use display::*;
 pub(crate) use platform::*;
-pub use scheduler::{Scheduler, TestScheduler, TestSchedulerConfig};
+pub use scheduler::{TestScheduler, TestSchedulerConfig};
 pub(crate) use window::*;
 
 pub use platform::TestScreenCaptureSource;
